[BACKPORT][wayland gpu] Gardening: move buffer management into own class.
This is a preliminary work before introducing frame and presentation
callbacks.

Summary: It's too much for WaylandConnection to hold that much
functionality. Instead, move the buffer management mechanism to its own
WaylandBufferManager class.

It's mostly moving code around with no functionality changes except
that CreateBuffer, SwapBuffer and DestroyBuffer return the result of
their job. On false, an error message is set and WaylandConnection
uses it when calling TerminateGpuProcess.

In followup CLs, I will make the buffer management easier.
That will allow us to use frame callbacks in quite a nice way.

Bug: 820047
Reviewed-on: https://chromium-review.googlesource.com/1179835
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#586201}
