+++ src/ui/ozone/platform/wayland/wayland_connection.cc
@@ -4,9 +4,6 @@
 
 #include "ui/ozone/platform/wayland/wayland_connection.h"
 
-#include <drm_fourcc.h>
-
-#include <linux-dmabuf-unstable-v1-client-protocol.h>
 #include <xdg-shell-unstable-v5-client-protocol.h>
 #include <xdg-shell-unstable-v6-client-protocol.h>
 
@@ -17,8 +14,7 @@
 #include "base/message_loop/message_loop_current.h"
 #include "base/strings/string_util.h"
 #include "base/threading/thread_task_runner_handle.h"
-#include "base/trace_event/trace_event.h"
-#include "ui/ozone/common/linux/drm_util_linux.h"
+#include "ui/ozone/platform/wayland/wayland_buffer_manager.h"
 #include "ui/ozone/platform/wayland/wayland_input_method_context.h"
 #include "ui/ozone/platform/wayland/wayland_object.h"
 #include "ui/ozone/platform/wayland/wayland_output_manager.h"
@@ -27,6 +23,7 @@
 static_assert(XDG_SHELL_VERSION_CURRENT == 5, "Unsupported xdg-shell version");
 
 namespace ui {
+
 namespace {
 const uint32_t kMaxCompositorVersion = 4;
 const uint32_t kMaxLinuxDmabufVersion = 1;
@@ -47,10 +44,7 @@ std::unique_ptr<WaylandDataSource> CreateWaylandDataSource(
 WaylandConnection::WaylandConnection()
     : controller_(FROM_HERE), binding_(this) {}
 
-WaylandConnection::~WaylandConnection() {
-  DCHECK(pending_buffer_map_.empty() && params_to_id_map_.empty() &&
-         buffers_.empty());
-}
+WaylandConnection::~WaylandConnection() = default;
 
 bool WaylandConnection::Initialize() {
   static const wl_registry_listener registry_listener = {
@@ -170,47 +164,6 @@ int WaylandConnection::GetKeyboardModifiers() {
   return modifiers;
 }
 
-// TODO(msisov): handle buffer swap failure or success.
-void WaylandConnection::ScheduleBufferSwap(const gfx::AcceleratedWidget& widget,
-                                           uint32_t buffer_id) {
-  DCHECK(base::MessageLoopForUI::IsCurrent());
-  if (!ValidateDataFromGpu(widget, buffer_id))
-    return;
-
-  auto it = buffers_.find(buffer_id);
-  // A buffer might not exist by this time. So, store the request and process
-  // it once it is created.
-  if (it == buffers_.end()) {
-    auto pending_buffers_it = pending_buffer_map_.find(buffer_id);
-    if (pending_buffers_it != pending_buffer_map_.end()) {
-      // If a buffer didn't exist and second call for a swap comes, buffer must
-      // be associated with the same widget.
-      DCHECK_EQ(pending_buffers_it->second, widget);
-    } else {
-      pending_buffer_map_.insert(
-          std::pair<uint32_t, gfx::AcceleratedWidget>(buffer_id, widget));
-    }
-    return;
-  }
-  struct wl_buffer* buffer = it->second.get();
-
-  WaylandWindow* window = GetWindow(widget);
-  if (!window)
-    return;
-
-  // TODO(msisov): it would be beneficial to use real damage regions to improve
-  // performance.
-  //
-  // TODO(msisov): also start using wl_surface_frame callbacks for better
-  // performance.
-  wl_surface_damage(window->surface(), 0, 0, window->GetBounds().width(),
-                    window->GetBounds().height());
-  wl_surface_attach(window->surface(), buffer, 0, 0);
-  wl_surface_commit(window->surface());
-
-  ScheduleFlush();
-}
-
 void WaylandConnection::CreateZwpLinuxDmabuf(
     base::File file,
     uint32_t width,
@@ -221,101 +174,31 @@ void WaylandConnection::CreateZwpLinuxDmabuf(
     const std::vector<uint64_t>& modifiers,
     uint32_t planes_count,
     uint32_t buffer_id) {
-  TRACE_EVENT2("Wayland", "WaylandConnection::CreateZwpLinuxDmabuf", "Format",
-               format, "Buffer id", buffer_id);
-
-  static const struct zwp_linux_buffer_params_v1_listener params_listener = {
-      WaylandConnection::CreateSucceeded, WaylandConnection::CreateFailed};
-
   DCHECK(base::MessageLoopForUI::IsCurrent());
-  if (!ValidateDataFromGpu(file, width, height, strides, offsets, format,
-                           modifiers, planes_count, buffer_id)) {
-    // base::File::Close() has an assertion that checks if blocking operations
-    // are allowed. Thus, manually close the fd here.
-    base::ScopedFD fd(file.TakePlatformFile());
-    fd.reset();
-    return;
+  if (!buffer_manager_->CreateBuffer(std::move(file), width, height, strides,
+                                     offsets, format, modifiers, planes_count,
+                                     buffer_id)) {
+    TerminateGpuProcess(buffer_manager_->error_message());
   }
-
-  // Store |params| connected to |buffer_id| to track buffer creation and
-  // identify, which buffer a client wants to use.
-  struct zwp_linux_buffer_params_v1* params =
-      zwp_linux_dmabuf_v1_create_params(zwp_linux_dmabuf());
-  params_to_id_map_.insert(
-      std::pair<struct zwp_linux_buffer_params_v1*, uint32_t>(params,
-                                                              buffer_id));
-  uint32_t fd = file.TakePlatformFile();
-  for (size_t i = 0; i < planes_count; i++) {
-    zwp_linux_buffer_params_v1_add(params, fd, i /* plane id */, offsets[i],
-                                   strides[i], 0 /* modifier hi */,
-                                   0 /* modifier lo */);
-  }
-  zwp_linux_buffer_params_v1_add_listener(params, &params_listener, this);
-  zwp_linux_buffer_params_v1_create(params, width, height, format, 0);
-
-  ScheduleFlush();
 }
 
 void WaylandConnection::DestroyZwpLinuxDmabuf(uint32_t buffer_id) {
-  TRACE_EVENT1("Wayland", "WaylandConnection::DestroyZwpLinuxDmabuf",
-               "Buffer id", buffer_id);
-
   DCHECK(base::MessageLoopForUI::IsCurrent());
-
-  auto it = buffers_.find(buffer_id);
-  if (it == buffers_.end()) {
-    TerminateGpuProcess("Trying to destroy non-existing buffer");
-    return;
+  if (!buffer_manager_->DestroyBuffer(buffer_id)) {
+    TerminateGpuProcess(buffer_manager_->error_message());
   }
-  buffers_.erase(it);
+}
 
-  ScheduleFlush();
+void WaylandConnection::ScheduleBufferSwap(const gfx::AcceleratedWidget& widget,
+                                           uint32_t buffer_id) {
+  DCHECK(base::MessageLoopForUI::IsCurrent());
+  if (!buffer_manager_->SwapBuffer(widget, buffer_id)) {
+    TerminateGpuProcess(buffer_manager_->error_message());
+  }
 }
 
 ClipboardDelegate* WaylandConnection::GetClipboardDelegate() {
   return this;
-}
-
-// static
-void WaylandConnection::CreateSucceeded(
-    void* data,
-    struct zwp_linux_buffer_params_v1* params,
-    struct wl_buffer* new_buffer) {
-  DCHECK(base::MessageLoopForUI::IsCurrent());
-
-  WaylandConnection* connection = static_cast<WaylandConnection*>(data);
-  DCHECK(connection);
-
-  // Find which buffer id |params| belong to and store wl_buffer
-  // with that id.
-  auto it = connection->params_to_id_map_.find(params);
-  CHECK(it != connection->params_to_id_map_.end());
-  uint32_t buffer_id = it->second;
-  connection->params_to_id_map_.erase(params);
-  zwp_linux_buffer_params_v1_destroy(params);
-
-  connection->buffers_.insert(std::pair<uint32_t, wl::Object<wl_buffer>>(
-      buffer_id, wl::Object<wl_buffer>(new_buffer)));
-
-  TRACE_EVENT1("Wayland", "WaylandConnection::CreateSucceeded", "Buffer id",
-               buffer_id);
-
-  auto pending_buffers_it = connection->pending_buffer_map_.find(buffer_id);
-  if (pending_buffers_it != connection->pending_buffer_map_.end()) {
-    gfx::AcceleratedWidget widget = pending_buffers_it->second;
-    connection->pending_buffer_map_.erase(pending_buffers_it);
-    connection->ScheduleBufferSwap(widget, buffer_id);
-  }
-}
-
-// static
-void WaylandConnection::CreateFailed(
-    void* data,
-    struct zwp_linux_buffer_params_v1* params) {
-  DCHECK(base::MessageLoopForUI::IsCurrent());
-
-  zwp_linux_buffer_params_v1_destroy(params);
-  LOG(FATAL) << "zwp_linux_buffer_params.create failed";
 }
 
 void WaylandConnection::OfferClipboardData(
@@ -347,14 +230,16 @@ bool WaylandConnection::IsSelectionOwner() {
 ozone::mojom::WaylandConnectionPtr WaylandConnection::BindInterface() {
   // This mustn't be called twice or when the zwp_linux_dmabuf interface is not
   // available.
-  DCHECK(!binding_.is_bound() || zwp_linux_dmabuf_);
+  DCHECK(!binding_.is_bound() || buffer_manager_);
   ozone::mojom::WaylandConnectionPtr ptr;
   binding_.Bind(MakeRequest(&ptr));
   return ptr;
 }
 
-std::vector<gfx::BufferFormat> WaylandConnection::GetSupportedBufferFormats() {
-  return buffer_formats_;
+const std::vector<gfx::BufferFormat>&
+WaylandConnection::GetSupportedBufferFormats() {
+  DCHECK(buffer_manager_);
+  return buffer_manager_->supported_buffer_formats();
 }
 
 void WaylandConnection::SetTerminateGpuCallback(
@@ -444,79 +329,10 @@ void WaylandConnection::OnFileCanReadWithoutBlocking(int fd) {
 
 void WaylandConnection::OnFileCanWriteWithoutBlocking(int fd) {}
 
-bool WaylandConnection::ValidateDataFromGpu(
-    const base::File& file,
-    uint32_t width,
-    uint32_t height,
-    const std::vector<uint32_t>& strides,
-    const std::vector<uint32_t>& offsets,
-    uint32_t format,
-    const std::vector<uint64_t>& modifiers,
-    uint32_t planes_count,
-    uint32_t buffer_id) {
-  std::string reason;
-  if (!file.IsValid())
-    reason = "Buffer fd is invalid";
-
-  if (width == 0 || height == 0)
-    reason = "Buffer size is invalid";
-
-  if (planes_count < 1)
-    reason = "Planes count cannot be less than 1";
-
-  if (planes_count != strides.size() || planes_count != offsets.size() ||
-      planes_count != modifiers.size()) {
-    reason = "Number of strides(" + std::to_string(strides.size()) +
-             ")/offsets(" + std::to_string(offsets.size()) + ")/modifiers(" +
-             std::to_string(modifiers.size()) +
-             ") does not correspond to the number of planes(" +
-             std::to_string(planes_count) + ")";
-  }
-
-  for (auto stride : strides) {
-    if (stride == 0)
-      reason = "Strides are invalid";
-  }
-
-  if (!IsValidBufferFormat(format))
-    reason = "Buffer format is invalid";
-
-  if (buffer_id < 1)
-    reason = "Invalid buffer id: " + std::to_string(buffer_id);
-
-  if (!reason.empty()) {
-    TerminateGpuProcess(reason);
-    return false;
-  }
-  return true;
-}
-
-bool WaylandConnection::ValidateDataFromGpu(
-    const gfx::AcceleratedWidget& widget,
-    uint32_t buffer_id) {
-  std::string reason;
-
-  if (widget == gfx::kNullAcceleratedWidget)
-    reason = "Invalid accelerated widget";
-
-  if (buffer_id < 1)
-    reason = "Invalid buffer id: " + std::to_string(buffer_id);
-
-  if (!reason.empty()) {
-    TerminateGpuProcess(reason);
-    return false;
-  }
-
-  return true;
-}
-
 void WaylandConnection::TerminateGpuProcess(std::string reason) {
   std::move(terminate_gpu_cb_).Run(std::move(reason));
   binding_.Unbind();
-
-  buffers_.clear();
-  params_to_id_map_.clear();
-  pending_buffer_map_.clear();
+  buffer_manager_->ClearState();
 }
 
 // static
@@ -533,9 +349,6 @@ void WaylandConnection::Global(void* data,
   };
   static const zxdg_shell_v6_listener shell_v6_listener = {
       &WaylandConnection::PingV6,
-  };
-  static const zwp_linux_dmabuf_v1_listener dmabuf_listener = {
-      &WaylandConnection::Format, &WaylandConnection::Modifiers,
   };
 
   WaylandConnection* connection = static_cast<WaylandConnection*>(data);
@@ -623,15 +436,13 @@ void WaylandConnection::Global(void* data,
     connection->data_device_manager_.reset(
         new WaylandDataDeviceManager(data_device_manager.release()));
     connection->data_device_manager_->set_connection(connection);
-  } else if (!connection->zwp_linux_dmabuf_ &&
+  } else if (!connection->buffer_manager_ &&
              (strcmp(interface, "zwp_linux_dmabuf_v1") == 0)) {
-    connection->zwp_linux_dmabuf_ = wl::Bind<zwp_linux_dmabuf_v1>(
-        registry, name, std::min(version, kMaxLinuxDmabufVersion));
-    zwp_linux_dmabuf_v1_add_listener(connection->zwp_linux_dmabuf(),
-                                     &dmabuf_listener, connection);
-    // A roundtrip after binding guarantees that the client has received all
-    // supported formats.
-    wl_display_roundtrip(connection->display_.get());
+    wl::Object<zwp_linux_dmabuf_v1> zwp_linux_dmabuf =
+        wl::Bind<zwp_linux_dmabuf_v1>(
+            registry, name, std::min(version, kMaxLinuxDmabufVersion));
+    connection->buffer_manager_.reset(
+        new WaylandBufferManager(zwp_linux_dmabuf.release(), connection));
   } else if (!connection->text_input_manager_v1_ &&
              strcmp(interface, "zwp_text_input_manager_v1") == 0) {
     connection->text_input_manager_v1_ = wl::Bind<zwp_text_input_manager_v1>(
@@ -730,27 +541,6 @@ void WaylandConnection::Ping(void* data, xdg_shell* shell, uint32_t serial) {
   WaylandConnection* connection = static_cast<WaylandConnection*>(data);
   xdg_shell_pong(shell, serial);
   connection->ScheduleFlush();
-}
-
-// static
-void WaylandConnection::Modifiers(void* data,
-                                  struct zwp_linux_dmabuf_v1* zwp_linux_dmabuf,
-                                  uint32_t format,
-                                  uint32_t modifier_hi,
-                                  uint32_t modifier_lo) {
-  NOTIMPLEMENTED();
-}
-
-// static
-void WaylandConnection::Format(void* data,
-                               struct zwp_linux_dmabuf_v1* zwp_linux_dmabuf,
-                               uint32_t format) {
-  WaylandConnection* connection = static_cast<WaylandConnection*>(data);
-  // Return on not the supported ARGB format.
-  if (format == DRM_FORMAT_ARGB2101010)
-    return;
-  connection->buffer_formats_.push_back(
-      GetBufferFormatFromFourCCFormat(format));
 }
 
 }  // namespace ui
