+++ src/ui/ozone/platform/wayland/wayland_connection.h
@@ -7,12 +7,11 @@
 
 #include <map>
 
-#include "ui/gfx/buffer_types.h"
-
 #include "base/files/file.h"
 #include "base/message_loop/message_pump_libevent.h"
 #include "mojo/public/cpp/bindings/binding.h"
 #include "ui/events/platform/platform_event_source.h"
+#include "ui/gfx/buffer_types.h"
 #include "ui/gfx/native_widget_types.h"
 #include "ui/ozone/platform/wayland/wayland_data_device.h"
 #include "ui/ozone/platform/wayland/wayland_data_device_manager.h"
@@ -23,16 +22,13 @@
 #include "ui/ozone/platform/wayland/wayland_pointer.h"
 #include "ui/ozone/platform/wayland/wayland_touch.h"
 #include "ui/ozone/public/clipboard_delegate.h"
-#include "ui/ozone/public/gpu_platform_support_host.h"
 #include "ui/ozone/public/interfaces/wayland/wayland_connection.mojom.h"
-
-struct zwp_linux_dmabuf_v1;
-struct zwp_linux_buffer_params_v1;
 
 namespace ui {
 
 class WaylandOutputManager;
 class WaylandWindow;
+class WaylandBufferManager;
 
 class WaylandConnection : public PlatformEventSource,
                           public ClipboardDelegate,
@@ -47,7 +43,7 @@ class WaylandConnection : public PlatformEventSource,
 
   // ozone::mojom::WaylandConnection overrides:
   //
-  // The overridden methods below are invoked by GPU.
+  // These overridden methods below are invoked by the GPU.
   //
   // Called by the GPU and asks to import a wl_buffer based on a gbm file
   // descriptor.
@@ -78,7 +74,6 @@ class WaylandConnection : public PlatformEventSource,
   zxdg_shell_v6* shell_v6() const { return shell_v6_.get(); }
   wl_seat* seat() { return seat_.get(); }
   wl_data_device* data_device() { return data_device_->data_device(); }
-  zwp_linux_dmabuf_v1* zwp_linux_dmabuf() { return zwp_linux_dmabuf_.get(); }
   zwp_text_input_manager_v1* text_input_manager_v1() {
     return text_input_manager_v1_.get();
   }
@@ -127,7 +122,7 @@ class WaylandConnection : public PlatformEventSource,
   // Returns bound pointer to own mojo interface.
   ozone::mojom::WaylandConnectionPtr BindInterface();
 
-  std::vector<gfx::BufferFormat> GetSupportedBufferFormats();
+  const std::vector<gfx::BufferFormat>& GetSupportedBufferFormats();
 
   void SetTerminateGpuCallback(
       base::OnceCallback<void(std::string)> terminate_gpu_cb);
@@ -170,19 +165,6 @@ class WaylandConnection : public PlatformEventSource,
   void OnFileCanReadWithoutBlocking(int fd) override;
   void OnFileCanWriteWithoutBlocking(int fd) override;
 
-  // Validates data sent by the GPU. If anything, terminates the gpu process.
-  bool ValidateDataFromGpu(const base::File& file,
-                           uint32_t width,
-                           uint32_t height,
-                           const std::vector<uint32_t>& strides,
-                           const std::vector<uint32_t>& offsets,
-                           uint32_t format,
-                           const std::vector<uint64_t>& modifiers,
-                           uint32_t planes_count,
-                           uint32_t buffer_id);
-  bool ValidateDataFromGpu(const gfx::AcceleratedWidget& widget,
-                           uint32_t buffer_id);
-
   // Terminates the GPU process on invalid data received
   void TerminateGpuProcess(std::string reason);
 
@@ -204,22 +186,6 @@ class WaylandConnection : public PlatformEventSource,
   // xdg_shell_listener
   static void Ping(void* data, xdg_shell* shell, uint32_t serial);
 
-  // zwp_linux_dmabuf_v1_listener
-  static void Modifiers(void* data,
-                        struct zwp_linux_dmabuf_v1* zwp_linux_dmabuf,
-                        uint32_t format,
-                        uint32_t modifier_hi,
-                        uint32_t modifier_lo);
-  static void Format(void* data,
-                     struct zwp_linux_dmabuf_v1* zwp_linux_dmabuf,
-                     uint32_t format);
-
-  static void CreateSucceeded(void* data,
-                              struct zwp_linux_buffer_params_v1* params,
-                              struct wl_buffer* new_buffer);
-  static void CreateFailed(void* data,
-                           struct zwp_linux_buffer_params_v1* params);
-
   std::map<gfx::AcceleratedWidget, WaylandWindow*> window_map_;
 
   wl::Object<wl_display> display_;
@@ -229,23 +195,8 @@ class WaylandConnection : public PlatformEventSource,
   wl::Object<wl_seat> seat_;
   wl::Object<wl_shm> shm_;
   wl::Object<xdg_shell> shell_;
-  wl::Object<zwp_linux_dmabuf_v1> zwp_linux_dmabuf_;
   wl::Object<zxdg_shell_v6> shell_v6_;
   wl::Object<zwp_text_input_manager_v1> text_input_manager_v1_;
-
-  // Stores a wl_buffer and it's id provided by the GbmBuffer object on the
-  // GPU process side.
-  base::flat_map<uint32_t, wl::Object<wl_buffer>> buffers_;
-  // A temporary params-to-buffer id map, which is used to identify which
-  // id wl_buffer should be assigned when storing it in the |buffers_| map
-  // during CreateSucceeded call.
-  base::flat_map<struct zwp_linux_buffer_params_v1*, uint32_t>
-      params_to_id_map_;
-  // It might happen that GPU asks to swap buffers, when a wl_buffer hasn't
-  // been created yet. Thus, store the request in a pending map. Once buffer
-  // is created, it will be attached to requested WaylandWindow based on the
-  // gfx::AcceleratedWidget.
-  base::flat_map<uint32_t, gfx::AcceleratedWidget> pending_buffer_map_;
 
   std::unique_ptr<WaylandDataDeviceManager> data_device_manager_;
   std::unique_ptr<WaylandDataDevice> data_device_;
@@ -255,6 +206,9 @@ class WaylandConnection : public PlatformEventSource,
   std::unique_ptr<WaylandOutputManager> wayland_output_manager_;
   std::unique_ptr<WaylandPointer> pointer_;
   std::unique_ptr<WaylandTouch> touch_;
+
+  // Objects that are using when GPU runs in own process.
+  std::unique_ptr<WaylandBufferManager> buffer_manager_;
 
   bool scheduled_flush_ = false;
   bool watching_ = false;
@@ -270,8 +224,6 @@ class WaylandConnection : public PlatformEventSource,
   RequestDataClosure read_clipboard_closure_;
 
   mojo::Binding<ozone::mojom::WaylandConnection> binding_;
-
-  std::vector<gfx::BufferFormat> buffer_formats_;
 
   // A callback, which is used to terminate a GPU process in case of invalid
   // data sent by the GPU to the browser process.
